// src/Model/iTechRestApi/iTechDataCaseStatusEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataCaseStatusEnum {
    notStarted = 1,
    workingOnIt = 2,
    processing = 3,
    stuckInternal = 4,
    waitingForAgentReview = 5,
    waitingForSupervisorReview = 6,
    onHold = 7,
    stuckExternal = 8,
    placeholder = 9,
    closed = 10
}