// src/Model/iTechRestApi/TableEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum TableEnum {
    iTechWebSim = 0,
    iTechWebAudit = 1,
    iTechWebSecurityObject = 2,
    iTechWebUser = 3,
    iTechStockHeader = 4,
    iTechStockOrderManagement = 5,
    iTechWebCaseManagement = 6,
    iTechWebTask = 7,
    iTechWebTaskOwner = 8,
    iTechWebSimAccident = 9,
    iTechWebSimSalesForce = 10,
    iTechWebTerm = 11,
    iTechWebSimUpload = 12,
    iTechStockQuote = 13,
    iTechWebUserGroup = 14,
    iTechWebRule = 15,
    iTechWebAlert = 16,
    iTechStockAimAudt = 17,
    iTechWebSurveillance = 18,
    iTechWebSurveillanceMonitor = 19,
    iTechWebLexicon = 20,
    iTechWebReportConfiguration = 21,
    iTechWebReport = 22,
    iTechWebLexiconGroup = 23,
    iTechWebHr = 24,
    iTechWebEodhdNews = 25,
    iTechWebStockOrder = 26,
    iTechWebManageEntity = 27,
    iTechWebCollectionServices = 28,
    iTechWebCollectionQueues = 29,
    iTechWebLicensing = 30,
    iTechWebReplication = 31,
    iTechWebHolderActions = 32,
    iTechWebCollectionTotals = 33,
    iTechWebReportingAssure = 34,
    iTechWebSimIpcUnigyRawCdr = 35,
    iTechWebSimNiceRawVoiceCdr = 36,
    iTechWebSimCiscoRawCdr = 37
}